+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # QOwnNotes Changelog
 
+## current
+- you now are able to select multiple notes and remove them or copy or move them to a recent note folder
+
 ## 0.39
 - turned on native file dialogs again
 - some text changes and svg image fixes